Refactor to shared receive component
