+++ src/components/receive.tsx
@@ -1,0 +1,82 @@
+import Row from "react-bootstrap/Row"
+import Col from "react-bootstrap/Col"
+import Card from "react-bootstrap/Card"
+import Container from "react-bootstrap/Container"
+import Image from "react-bootstrap/Image"
+
+import { getOS, appStoreLink, playStoreLink } from "./downloadApp"
+import ReceiveAmount from "./receiveAmount"
+import ReceiveNoAmount from "./receiveNoAmount"
+
+export default function Receive({
+  username,
+  amount,
+}: {
+  username: string
+  amount?: number
+}) {
+  const os = getOS()
+
+  return (
+    <Container fluid>
+      {os === undefined && <br />}
+      <Row className="justify-content-md-center">
+        <Col md="auto" style={{ padding: 0 }}>
+          <Card className="text-center">
+            {amount ? (
+              <ReceiveAmount username={username} amount={amount} />
+            ) : (
+              <ReceiveNoAmount username={username} />
+            )}
+
+            <Card.Body>
+              {os === "android" && (
+                <a href={playStoreLink}>
+                  <Image
+                    src={process.env.PUBLIC_URL + "/google-play-badge.png"}
+                    height="40px"
+                    rounded
+                  />
+                </a>
+              )}
+              {os === "ios" && (
+                <a href={playStoreLink}>
+                  <Image
+                    src={process.env.PUBLIC_URL + "/apple-app-store.png"}
+                    height="40px"
+                    rounded
+                  />
+                </a>
+              )}
+              {os === undefined && (
+                <div>
+                  <a href={appStoreLink}>
+                    <Image
+                      src={process.env.PUBLIC_URL + "/apple-app-store.png"}
+                      height="45px"
+                      rounded
+                    />
+                  </a>
+                  &nbsp;
+                  <a href={playStoreLink}>
+                    <Image
+                      src={process.env.PUBLIC_URL + "/google-play-badge.png"}
+                      height="45px"
+                      rounded
+                    />
+                  </a>
+                </div>
+              )}
+            </Card.Body>
+            <Card.Footer className="text-muted">
+              Powered by <Card.Link href="https://galoy.io">Galoy </Card.Link>
+              <br />
+              <Card.Link href={window.location.origin}>Open a channel with us</Card.Link>
+            </Card.Footer>
+          </Card>
+        </Col>
+      </Row>
+      <br />
+    </Container>
+  )
+}
